isorl: fix compilation

+++ examples/isorl/src/area/area.c
@@ -155,7 +155,7 @@ Area *area_load(World *w, uint16_t id)
    RvR_rw_close(&rw);
 
    RvR_rw_init_const_mem(&rw_decomp, area_buffer, size);
-   int32_t size_out = 0;
+   size_t size_out = 0;
    mem_decomp = RvR_crush_decompress(&rw_decomp, &size_out);
    RvR_mem_tag_set(mem_decomp, RVR_MALLOC_STATIC);
    RvR_rw_close(&rw_decomp);
